# lib/api-zod/src/generated/types/adminWithdrawal.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * StarCash Affiliate Earning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminWithdrawalStatus } from './adminWithdrawalStatus';

export interface AdminWithdrawal {
  id: number;
  userId: number;
  username: string;
  fullName: string;
  bankName: string;
  accountName: string;
  accountNumber: string;
  amount: number;
  status: AdminWithdrawalStatus;
  createdAt: Date;
  updatedAt?: Date;
}